Validate Permissions interface.
